refactor: subscriptions tests to their own file
